[Silver II] Title: 생태학, Time: 288 ms, Memory: 2292 KB -BaekjoonHub

+++ 백준/Silver/4358. 생태학/생태학.cc
@@ -5,11 +5,11 @@ int main(){
     ios_base::sync_with_stdio(0);
     cin.tie(0);
     cout.tie(0);
-
+    
     map<string, int> trees;
     string str;
 
-    float cnt;
+    double cnt;
 
     while(getline(cin, str)){
         cnt++;
